Updated CLTests to change how we think about add(int index, E item)
You can no longer use add(int index, E item) to add to the end of the list.  Period.
Instead, we shift every element over and add to index = 0.  This seems to fit
better with the overall idea of the CircularList, and does not remove functionallity.

If you wish to add to the end of the list, there is add(E item);